+++ Tests/Unit/Controller/BackEnd/RegistrationControllerTest.php
@@ -217,6 +217,9 @@ final class RegistrationControllerTest extends UnitTestCase
         $this->registrationRepositoryMock
             ->expects(self::never())->method('findWaitingListRegistrationsByEvent')
             ->with(self::anything());
+        $this->registrationRepositoryMock
+            ->expects(self::never())->method('findNonbindingReservationsByEvent')
+            ->with(self::anything());
 
         $this->subject->showForEventAction($eventUid);
     }
@@ -249,13 +252,14 @@ final class RegistrationControllerTest extends UnitTestCase
             ->method('findOneByUidForBackend')->with($eventUid)->willReturn($event);
 
         $this->viewMock
-            ->expects(self::exactly(5))->method('assign')
+            ->expects(self::exactly(6))->method('assign')
             ->withConsecutive(
                 ['permissions', $this->permissionsMock],
                 ['pageUid', self::anything()],
                 ['event', self::anything()],
                 ['regularRegistrations', self::anything()],
                 ['waitingListRegistrations', self::anything()],
+                ['nonbindingReservations', self::anything()],
             );
 
         $this->subject->showForEventAction($eventUid);
@@ -275,13 +279,14 @@ final class RegistrationControllerTest extends UnitTestCase
             ->method('findOneByUidForBackend')->with($eventUid)->willReturn($event);
 
         $this->viewMock
-            ->expects(self::exactly(5))->method('assign')
+            ->expects(self::exactly(6))->method('assign')
             ->withConsecutive(
                 ['permissions', self::anything()],
                 ['pageUid', $pageUid],
                 ['event', self::anything()],
                 ['regularRegistrations', self::anything()],
                 ['waitingListRegistrations', self::anything()],
+                ['nonbindingReservations', self::anything()],
             );
 
         $this->subject->showForEventAction($eventUid);
@@ -299,13 +304,14 @@ final class RegistrationControllerTest extends UnitTestCase
             ->method('findOneByUidForBackend')->with($eventUid)->willReturn($event);
 
         $this->viewMock
-            ->expects(self::exactly(5))->method('assign')
+            ->expects(self::exactly(6))->method('assign')
             ->withConsecutive(
                 ['permissions', self::anything()],
                 ['pageUid', self::anything()],
                 ['event', $event],
                 ['regularRegistrations', self::anything()],
                 ['waitingListRegistrations', self::anything()],
+                ['nonbindingReservations', self::anything()],
             );
 
         $this->subject->showForEventAction($eventUid);
@@ -345,13 +351,14 @@ final class RegistrationControllerTest extends UnitTestCase
             ->with($eventUid)->willReturn($regularRegistrations);
 
         $this->viewMock
-            ->expects(self::exactly(5))->method('assign')
+            ->expects(self::exactly(6))->method('assign')
             ->withConsecutive(
                 ['permissions', self::anything()],
                 ['pageUid', self::anything()],
                 ['event', self::anything()],
                 ['regularRegistrations', $regularRegistrations],
                 ['waitingListRegistrations', self::anything()],
+                ['nonbindingReservations', self::anything()],
             );
 
         $this->subject->showForEventAction($eventUid);
@@ -373,8 +380,8 @@ final class RegistrationControllerTest extends UnitTestCase
             ->expects(self::once())->method('findRegularRegistrationsByEvent')
             ->with($eventUid)->willReturn($regularRegistrations);
         $this->registrationRepositoryMock
-            ->expects(self::exactly(2))->method('enrichWithRawData')
-            ->withConsecutive([$regularRegistrations], [self::anything()]);
+            ->expects(self::exactly(3))->method('enrichWithRawData')
+            ->withConsecutive([$regularRegistrations], [self::anything()], [self::anything()]);
 
         $this->subject->showForEventAction($eventUid);
     }
@@ -397,13 +404,14 @@ final class RegistrationControllerTest extends UnitTestCase
             ->with($eventUid)->willReturn($waitingListRegistrations);
 
         $this->viewMock
-            ->expects(self::exactly(5))->method('assign')
+            ->expects(self::exactly(6))->method('assign')
             ->withConsecutive(
                 ['permissions', self::anything()],
                 ['pageUid', self::anything()],
                 ['event', self::anything()],
                 ['regularRegistrations', self::anything()],
                 ['waitingListRegistrations', $waitingListRegistrations],
+                ['nonbindingReservations', self::anything()],
             );
 
         $this->subject->showForEventAction($eventUid);
@@ -427,13 +435,14 @@ final class RegistrationControllerTest extends UnitTestCase
             ->with($eventUid)->willReturn($waitingListRegistrations);
 
         $this->viewMock
-            ->expects(self::exactly(5))->method('assign')
+            ->expects(self::exactly(6))->method('assign')
             ->withConsecutive(
                 ['permissions', self::anything()],
                 ['pageUid', self::anything()],
                 ['event', self::anything()],
                 ['regularRegistrations', self::anything()],
                 ['waitingListRegistrations', $waitingListRegistrations],
+                ['nonbindingReservations', self::anything()],
             );
 
         $this->subject->showForEventAction($eventUid);
@@ -456,8 +465,8 @@ final class RegistrationControllerTest extends UnitTestCase
             ->expects(self::once())->method('findWaitingListRegistrationsByEvent')
             ->with($eventUid)->willReturn($waitingListRegistrations);
         $this->registrationRepositoryMock
-            ->expects(self::exactly(2))->method('enrichWithRawData')
-            ->withConsecutive([self::anything()], [$waitingListRegistrations]);
+            ->expects(self::exactly(3))->method('enrichWithRawData')
+            ->withConsecutive([self::anything()], [$waitingListRegistrations], [self::anything()]);
 
         $this->subject->showForEventAction($eventUid);
     }
@@ -479,8 +488,60 @@ final class RegistrationControllerTest extends UnitTestCase
             ->expects(self::once())->method('findWaitingListRegistrationsByEvent')
             ->with($eventUid)->willReturn($waitingListRegistrations);
         $this->registrationRepositoryMock
-            ->expects(self::exactly(2))->method('enrichWithRawData')
-            ->withConsecutive([self::anything()], [$waitingListRegistrations]);
+            ->expects(self::exactly(3))->method('enrichWithRawData')
+            ->withConsecutive([self::anything()], [$waitingListRegistrations], [self::anything()]);
+
+        $this->subject->showForEventAction($eventUid);
+    }
+
+    /**
+     * @test
+     */
+    public function showForEventActionEnrichesNonbindingReservationsWithRawData(): void
+    {
+        $eventUid = 5;
+        $event = $this->buildSingleEventMockWithUid($eventUid);
+        $this->eventRepositoryMock
+            ->expects(self::once())
+            ->method('findOneByUidForBackend')->with($eventUid)->willReturn($event);
+        $nonbindingReservations = [new Registration()];
+
+        $this->registrationRepositoryMock
+            ->expects(self::once())->method('findNonbindingReservationsByEvent')
+            ->with($eventUid)->willReturn($nonbindingReservations);
+        $this->registrationRepositoryMock
+            ->expects(self::exactly(3))->method('enrichWithRawData')
+            ->withConsecutive([self::anything()], [self::anything()], [$nonbindingReservations]);
+
+        $this->subject->showForEventAction($eventUid);
+    }
+
+    /**
+     * @test
+     */
+    public function showForEventActionPassesNonbindingReservationsForProvidedEventToView(): void
+    {
+        $eventUid = 5;
+        $event = $this->buildSingleEventMockWithUid($eventUid);
+        $this->eventRepositoryMock
+            ->expects(self::once())
+            ->method('findOneByUidForBackend')->with($eventUid)->willReturn($event);
+        $nonbindingReservations = [new Registration()];
+
+        $this->registrationRepositoryMock
+            ->expects(self::once())->method('findNonbindingReservationsByEvent')
+            ->with($eventUid)->willReturn($nonbindingReservations);
+
+        $this->viewMock
+            ->expects(self::exactly(6))->method('assign')
+            ->withConsecutive(
+                ['permissions', self::anything()],
+                ['pageUid', self::anything()],
+                ['event', self::anything()],
+                ['regularRegistrations', self::anything()],
+                ['waitingListRegistrations', self::anything()],
+                ['nonbindingReservations', $nonbindingReservations],
+            );
 
         $this->subject->showForEventAction($eventUid);
     }